# accounting/openapi/openapi.json
{
...
  "host": "127.0.0.1:45881",
  "info": {
    "title": "Komunitin accounting API",
    "version": "0.0.1",
    "license": {
      "name": "AGPL-3.0-only"
    },
    "description": "API for the Komunitin payments system. This documentation is automatically generated from the serivce test suite."
  },
  "paths": {
...
    "/{code}/creditCommonsNodes": {
      "post": {
        "summary": "/{code}/creditCommonsNodes",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "type": "string",
            "example": "TEST"
          },
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "object",
                  "properties": {
                    "attributes": {
                      "type": "object",
                      "properties": {
                        "ccNodeName": {
                          "type": "string",
                          "example": "trunk"
                        },
                        "routingPrefix": {
                          "type": "string",
                          "example": "trunk/branch2"
                        },
                        "lastHash": {
                          "type": "string",
                          "example": "asdf"
                        },
                        "vostroId": {
                          "type": "string",
                          "example": "77978725-368e-41a1-973c-4115440dfb48"
                        }
                      }
                    },
                    "relationships": {
                      "type": "object",
                      "properties": {
                        "vostro": {
                          "type": "object",
                          "properties": {
                            "data": {
                              "type": "object",
                              "properties": {
                                "type": {
                                  "type": "string",
                                  "example": "accounts"
                                },
                                "id": {
                                  "type": "string",
                                  "example": "77978725-368e-41a1-973c-4115440dfb48"
                                }
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string",
                      "example": "creditCommonsNodes"
                    },
                    "attributes": {
                      "type": "object",
                      "properties": {
                        "ccNodeName": {
                          "type": "string",
                          "example": "trunk"
                        },
                        "lastHash": {
                          "type": "string",
                          "example": "asdf"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "type": "object",
              "properties": {
                "errors": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "status": {
                        "type": "string"
                      },
                      "code": {
                        "type": "string"
                      },
                      "title": {
                        "type": "string"
                      },
                      "detail": {
                        "type": "string"
                      }
                    }
                  },
                  "example": [
                    {
                      "status": "400",
                      "code": "FieldValidationError",
                      "title": "Field validation Error",
                      "detail": "Field data.attributes.routingPrefix is invalid in request body."
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden",
            "schema": {
              "type": "object",
              "properties": {
                "errors": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "status": {
                        "type": "string"
                      },
                      "code": {
                        "type": "string"
                      },
                      "title": {
                        "type": "string"
                      },
                      "detail": {
                        "type": "string"
                      }
                    }
                  },
                  "example": [
                    {
                      "status": "403",
                      "code": "Forbidden",
                      "title": "Forbidden",
                      "detail": "Insufficient Scope"
                    }
                  ]
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "type": "object",
              "properties": {
                "errors": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "status": {
                        "type": "string"
                      },
                      "code": {
                        "type": "string"
                      },
                      "title": {
                        "type": "string"
                      },
                      "detail": {
                        "type": "string"
                      }
                    }
                  },
                  "example": [
                    {
                      "status": "500",
                      "code": "InternalError",
                      "title": "Internal Error",
                      "detail": "\nInvalid `this.db().creditCommonsNode.create()` invocation in\n/Users/michiel/gh/komunitin/komunitin/accounting/src/creditcommons/credit-commons-controller.ts:95:39\n\n  92 async createNode(ctx: Context, ccNodeName: string, lastHash: string, vostroId: string): Promise<CreditCommonsNode> {\n  93   // Only admins are allowed to set the trunkward node:\n  94   await this.users().checkAdmin(ctx)\n→ 95   await this.db().creditCommonsNode.create({\n         data: {\n           tenantId: \"TEST\",\n           ccNodeName: \"trunk\",\n           lastHash: \"asdf\",\n           vostroId: \"a9fb66cf-468a-482d-a0b5-14d0d0e98539\",\n       +   routingPrefix: String\n         }\n       })\n\nArgument `routingPrefix` is missing."
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [],
        "produces": [
          "application/vnd.api+json"
        ],
        "security": [
          {
            "authorization": []
          }
        ]
      }
    },
    "/{code}/cc/": {
      "get": {
        "summary": "/{code}/cc/",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true
          }
        ],
        "responses": {},
        "tags": []
      }
    },
    "/{code}/cc/transaction/relay": {
      "post": {
        "summary": "/{code}/cc/transaction/relay",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true
          }
        ],
        "responses": {},
        "tags": []
      }
    },
    "/{code}/cc/transaction/{transId}/{newState}": {
      "patch": {
        "summary": "/{code}/cc/transaction/{transId}/{newState}",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true
          },
          {
            "name": "transId",
            "in": "path",
            "required": true
          },
          {
            "name": "newState",
            "in": "path",
            "required": true
          }
        ],
        "responses": {},
        "tags": []
      }
    },
    "/{code}/cc/account": {
      "get": {
        "summary": "/{code}/cc/account",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true
          }
        ],
        "responses": {},
        "tags": []
      }
    },
    "/{code}/cc/account/history": {
      "get": {
        "summary": "/{code}/cc/account/history",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true
          }
        ],
        "responses": {},
        "tags": []
      }
    },
...
  },
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "authorization": {
      "name": "authorization",
      "in": "header",
      "type": "apiKey"
    }
  },
  "servers": [
    {
      "url": "https://accounting.komunitin.org",
      "description": "Production server"
    },
    {
      "url": "https://accounting.demo.komunitin.org",
      "description": "Demo server"
    },
    {
      "url": "http://localhost:2025",
      "description": "Local development server"
    }
  ],
  "swagger": "2.0",
...
}